Fixed bug in GfxStreamLogger related to DrawLineStraight.
(cherry picked from commit bc974922111db3018487361d1f5fc6f98604b9ff)

+++ src/gfxdevices/stream/wg_gfxstreamlogger.cpp
@@ -329,7 +329,7 @@ namespace wg
 
 				CoordI		begin;
 				Direction	dir;
-				uint16_t	length;
+				int			length;
 				HiColor		color;
 				float		thickness;
